+++ test/AElf.Kernel.SmartContract.ExecutionPluginForCallThreshold.Tests/MockCoefficientProvider.cs
@@ -50,7 +50,7 @@ namespace AElf.Kernel.SmartContract.ExecutionPluginForCallThreshold.Tests
                 return Task.FromResult(_coefficientsDicCache[tokenType]);
             }
 
-            public void SetModifyHeight(int tokenType)
+            public void SetModifyHeight(long tokenType)
             {
             }
             public Task SyncCacheAsync(IChainContext chainContext)
