+++ test/AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests/MockCoefficientProvider.cs
@@ -49,7 +49,7 @@ namespace AElf.Kernel.SmartContract.ExecutionPluginForMethodFee.Tests
         {
             return Task.FromResult(_coefficientsDicCache[tokenType]);
         }
-        public void SetModifyHeight(int tokenType)
+        public void SetModifyHeight(long tokenType)
         {
         }
         public Task SyncCacheAsync(IChainContext chainContext)
